Use fully-qualified iconv calls in the mbstring polyfill
The mbstring polyfill does not require the iconv polyfill due to the
size of the package and the fact that most systems will have the native
ext-iconv (it is enabled by default in PHP).
Using fully qualified calls gives a better error message when the iconv
extension is missing and the polyfill is not installed, which makes it
easier to solve the issue.