Fix QQQQQ date format
As per [1], the QQQQQ and qqqqq quarter patterns are supposed to convert
quarter into numbers (1, 2, 3, or 4). Until vertion 70.1, ICU carried a
bug [2] which would make those patterns to convert into the wrong string,
mimicking the behavior of the QQQQ pattern. The issue has been fixed
upstream [3] and released in ICU 70.1.

This patch adjusts the pattern resolution according to the ICU version
being used by intl.

[1] https://unicode-org.github.io/icu/userguide/format_parse/datetime/#date-field-symbol-table
[2] https://unicode-org.atlassian.net/browse/ICU-21647
[3] https://github.com/unicode-org/icu/commit/dcfdaca46c1e3cc08764ad1d85f321ad0c27ae16

Signed-off-by: Athos Ribeiro <athos.ribeiro@canonical.com>