[PHP 8.5] Add `array_first` and `array_last`
Adds polyfills for the PHP 8.5's `array_first` and `array_last`
functions.

The polyfills would have been a bit cleaner if we used the
`array_key_first` and `array_key_last` functions, but they were added
in PHP 7.3, and are not available on PHP 7.2.

This adds the same tests from php-src patch.

 - [RFC: `array_first()` and `array_last()`](https://wiki.php.net/rfc/array_first_last)
 - [php-src commit](https://github.com/php/php-src/commit/168343d2e811fe0f2152d39a8f88557cc0b35c86)
 - [PHP.Watch polyfill for PHP 8.0+](https://php.watch/versions/8.5/array_first-array_last)

Co-authored-by: Nicolas Grekas <nicolas.grekas@gmail.com>